+++ src/api/services/answer/common/useAnswerQuestion.ts
@@ -1,0 +1,13 @@
+import { useMutation } from '@tanstack/react-query'
+
+import { answerRandomQuestion } from '.'
+
+type AnswerQuestionParam = {
+  friendId: number
+}
+
+export const useAnswerQuestion = ({ friendId }: AnswerQuestionParam) => {
+  return useMutation({
+    mutationFn: () => answerRandomQuestion({ friendId }),
+  })
+}
