+++ src/api/services/friend/useKakaoFriends.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+
+import { fetchKakaoFriends } from '.'
+
+export const useKakaoFriends = () => {
+  return useQuery({
+    queryKey: ['kakaoFriends'],
+    queryFn: fetchKakaoFriends,
+  })
+}
